backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
pymysql
cryptography
